requirements.txt
streamlit
python-docx
PyPDF2
sentence-transformers
scikit-learn
numpy
requests 